+++ CHANGELOG.md
@@ -5,6 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.3] - 2023-06-20
+
+### Added
+
+- **Crypto**: Hex output when the size matches a classic standard.
+- **Native**: `UUID` support for hex format.
+
+### Fixed
+
+- **Crypto**: Fixed display of `useKeyGen.build`.
+
+### Changed
+
+- **Crypto**: Better error support when coding with null values.
+
 ## [1.0.2] - 2023-06-16
 
 ### Fixed
@@ -23,6 +38,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 - Initial Release.
 
+[1.0.3]: https://github.com/hyugogirubato/Frida-CodeShare/releases/tag/v1.0.3
 [1.0.2]: https://github.com/hyugogirubato/Frida-CodeShare/releases/tag/v1.0.2
 [1.0.1]: https://github.com/hyugogirubato/Frida-CodeShare/releases/tag/v1.0.1
 [1.0.0]: https://github.com/hyugogirubato/Frida-CodeShare/releases/tag/v1.0.0